+++ cobras/runner/subscribe.py
@@ -14,11 +14,11 @@ import uvloop
 from cobras.client.client import subscribeClient
 from cobras.client.credentials import (createCredentials, getDefaultRoleForApp,
                                        getDefaultSecretForApp)
-from cobras.common.apps_config import PUBSUB_APPKEY
+from cobras.common.apps_config import PUBSUB_APPKEY, getDefaultPort
 from cobras.common.superuser import preventRootUsage
 from cobras.common.throttle import Throttle
 
-DEFAULT_URL = f'ws://127.0.0.1:8765/v2?appkey={PUBSUB_APPKEY}'
+DEFAULT_URL = f'ws://127.0.0.1:{getDefaultPort()}/v2?appkey={PUBSUB_APPKEY}'
 
 
 class MessageHandlerClass: