+++ cobras/runner/write.py
@@ -12,10 +12,10 @@ import uvloop
 from cobras.client.connection import Connection
 from cobras.client.credentials import (createCredentials, getDefaultRoleForApp,
                                        getDefaultSecretForApp)
-from cobras.common.apps_config import PUBSUB_APPKEY
+from cobras.common.apps_config import PUBSUB_APPKEY, getDefaultPort
 from cobras.common.superuser import preventRootUsage
 
-DEFAULT_URL = f'ws://127.0.0.1:8765/v2?appkey={PUBSUB_APPKEY}'
+DEFAULT_URL = f'ws://127.0.0.1:{getDefaultPort()}/v2?appkey={PUBSUB_APPKEY}'
 
 
 @click.command()
@@ -34,7 +34,7 @@ def write(url, role, secret, channel, data):
     credentials = createCredentials(role, secret)
 
     async def handler(url, credentials, channel, data):
-        connection = Connection(url, credentials, verbose=True)
+        connection = Connection(url, credentials)
         await connection.connect()
         await connection.write(channel, data)
         await connection.close()